requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
data-wizard==2.0.0
diff-match-patch==20230430
distlib==0.3.8
Django==5.0.4
django-import-export==4.0.6
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
et-xmlfile==1.1.0
filelock==3.13.1
html-json-forms==1.1.1
idna==3.7
iniconfig==2.0.0
itertable==2.2.0
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.3.0
mysqlclient==2.2.4
natural-keys==2.1.1
openpyxl==3.1.2
packaging==23.2
platformdirs==4.2.0
pluggy==1.4.0
pytest==8.0.1
python-dateutil==2.9.0.post0
requests==2.32.2
six==1.16.0
sqlparse==0.5.0
tablib==3.5.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.0
